fix: split Development screen body under 400-line screen cap (#4175)
Extract region tab and assign handler so repo.app_features_screens_lib_physical_file_size passes after #4305 ratchet landed on dev.
